wexam/requirements-docker.txt
bcrypt==3.1.4
cffi==1.11.5
click==6.7
dectate==0.13
importscan==0.1
more.jwtauth==0.11
more.pony==0.2
morepath==0.18.1
passlib==1.7.1
pony==0.7.3
psycopg2-binary==2.7.5
pycparser==2.18
PyJWT==1.5.3
PyYAML==3.12
redis==2.10.6
reg==0.11
repoze.lru==0.7
rq==0.11.0
simhash==1.9.0
six==1.11.0
uWSGI==2.0.17
WebOb==1.8.2
Werkzeug==0.14.1
